fix contact apis
fix contact apis: update, delete

+++ controllers/contact.controller.js
@@ -67,6 +67,10 @@ const updateContact = asyncHandler(async (req, res) => {
         res.status(404);     
         throw new Error("Contact not found");
     }
+    if (constact.user_id.toString() !== req.user.id) {
+        res.status(403);
+        throw new Error("User don't have permission to update other user contacts");
+    }
     const updatedContact = await Contact.findByIdAndUpdate(
         req.params.id,
         req.body,
@@ -87,6 +91,10 @@ const deleteContact = asyncHandler(async (req, res) => {
     if (!constact) {
         res.status(404);     
         throw new Error("Contact not found");
+    }
+    if (constact.user_id.toString() !== req.user.id) {
+        res.status(403);
+        throw new Error("User don't have permission to delete other user contacts");
     }
     await Contact.deleteOne({ _id: req.params.id });
     // await Contact.rem
